rewrote some functions from `scene.js` in rust
I should try to write as little as possible in rust, so I probably won't end up
needing the new functions in `lib.rs`. I should store the scene and other
important data in wasm's linear memory to avoid the overhead of copying data
between wasm memory and the JS heap. This will make it easier for `render()` to
draw to the canvas (from rust), but it will also mean that `makeSphere()`,
`deleteSphere()` and co. (in JS) will have to deal with wasm memory.