removed web_sys from Rust
it's much easier to write to a buffer in Rust then access wasm memory from JS
and copy it to the canvas. Also switched from a Vec<u32> buffer to a Vec<u8>
that's four times as large for the framebuffer. Vec<u32> only showed black and
white circles, but switching to Vec<u8> seems to have broken something since
nothing shows up now.